+++ src/dao/impl/TicketRepositoryImpl.java
@@ -3,7 +3,7 @@ package dao.impl;
 import dao.repository.TicketRepository;
 import model.entity.Ticket;
 
-public class TicketRepositoryImpl extends AbstractRepository<Long, Ticket> implements TicketRepository {
+public class TicketRepositoryImpl extends AbstractRepository<Integer, Ticket> implements TicketRepository {
     public TicketRepositoryImpl(IdGenerator idGenerator) {
         super(idGenerator);
     }
