+++ src/dao/impl/WorkerRepositoryImpl.java
@@ -3,7 +3,7 @@ package dao.impl;
 import dao.repository.WorkerRepository;
 
 public class WorkerRepositoryImpl extends UserRepositoryImpl implements WorkerRepository {
-    public WorkerRepositoryImpl(IdGenerator<Long> idGenerator) {
+    public WorkerRepositoryImpl(IdGenerator<Integer> idGenerator) {
         super(idGenerator);
     }
 }
