+++ src/dao/repository/MovieRepository.java
@@ -7,7 +7,7 @@ import model.entity.Movie;
 import java.time.LocalDate;
 import java.util.Collection;
 
-public interface MovieRepository extends Repository<Long, Movie>{
+public interface MovieRepository extends Repository<Integer, Movie>{
     Movie findByMovieName(String movieName) throws NonExistingEntityException;
     Collection<Movie> findByPrice(double moviePrice) throws NonExistingEntityException;
     Collection<Movie> findByGenre(Genre genre) throws NonExistingEntityException;
