+++ src/dao/repository/UserRepository.java
@@ -4,7 +4,7 @@ import dao.exception.NonExistingEntityException;
 import model.UserStatus;
 import model.entity.User;
 
-public interface UserRepository extends Repository<Long, User> {
+public interface UserRepository extends Repository<Integer, User> {
     User findByUsername(String username) throws NonExistingEntityException;
     User findByEmail(String email) throws NonExistingEntityException;
